refactor: centralize common layout components
