+++ src/cnd-app/for-drivers/index.jsx
@@ -1,13 +1,13 @@
 "use client";
 
 import React from "react";
-import { Navbar12 } from "../home/components/Navbar12";
+import { Navbar12 } from "../../components/common/Navbar12";
 import { Header47 } from "./components/Header47";
 import { Layout3 } from "./components/Layout3";
 import { Layout12 } from "./components/Layout12";
 import { Layout3_1 } from "./components/Layout3_1";
 import { Cta1 } from "./components/Cta1";
-import { Footer2 } from "../home/components/Footer2";
+import { Footer2 } from "../../components/common/Footer2";
 
 export default function ForDrivers() {
   return (
