+++ src/cnd-app/for-hosts/index.jsx
@@ -1,14 +1,14 @@
 "use client";
 
 import React from "react";
-import { Navbar12 } from "../home/components/Navbar12";
+import { Navbar12 } from "../../components/common/Navbar12";
 import { Header47 } from "./components/Header47";
 import { Layout6 } from "./components/Layout6";
 import { Layout242 } from "./components/Layout242";
 import { Layout27 } from "./components/Layout27";
 import { Testimonial6 } from "./components/Testimonial6";
 import { Cta19 } from "./components/Cta19";
-import { Footer2 } from "../home/components/Footer2";
+import { Footer2 } from "../../components/common/Footer2";
 
 export default function ForHosts() {
   return (
