+++ src/cnd-app/home/index.jsx
@@ -1,13 +1,13 @@
 "use client";
 
 import React from "react";
-import { Navbar12 } from "./components/Navbar12";
+import { Navbar12 } from "../../components/common/Navbar12";
 import { Header1 } from "./components/Header1";
 import { Section2 } from "./components/section2";
 import { Testimonial14 } from "./components/Testimonial14";
 import { Cta7 } from "./components/Cta7";
 import { Cta20 } from "./components/Cta20";
-import { Footer2 } from "./components/Footer2";
+import { Footer2 } from "../../components/common/Footer2";
 
 export default function Home() {
   return (
